requirements.txt
torch==2.5.0
torchvision==0.15.2
pandas==1.5.3
numpy==1.23.5
Pillow==11.1.0
mlflow==2.16.0
pytest==7.2.1
PyYAML~=5.4.1
scikit-learn~=1.5.2
mlflow-skinny~=2.16.0
DateTime~=5.4
networkx~=3.2.1
tqdm~=4.66.1
tabulate~=0.9.0
optuna~=4.2.0